add docdb module, add read from docdb to load previous alignments. https://github.com/AllenNeuralDynamics/aind-physio-arch/issues/708
